requirements.txt
Flask==3.0.2
openai==1.3.7
pdfplumber==0.10.3
python-dotenv==1.0.1
flask-cors==4.0.0
requests==2.31.0
